refactor: split clipboard module implementation for browser / renderer
