+++ filenames.auto.gni
@@ -196,6 +196,7 @@ auto_filenames = {
     "lib/browser/api/base-window.ts",
     "lib/browser/api/browser-view.ts",
     "lib/browser/api/browser-window.ts",
+    "lib/browser/api/clipboard.ts",
     "lib/browser/api/content-tracing.ts",
     "lib/browser/api/crash-reporter.ts",
     "lib/browser/api/desktop-capturer.ts",
@@ -245,7 +246,6 @@ auto_filenames = {
     "lib/browser/parse-features-string.ts",
     "lib/browser/rpc-server.ts",
     "lib/browser/web-view-events.ts",
-    "lib/common/api/clipboard.ts",
     "lib/common/api/module-list.ts",
     "lib/common/api/native-image.ts",
     "lib/common/api/shell.ts",
@@ -256,8 +256,6 @@ auto_filenames = {
     "lib/common/reset-search-paths.ts",
     "lib/common/web-view-methods.ts",
     "lib/common/webpack-globals-provider.ts",
-    "lib/renderer/ipc-renderer-internal-utils.ts",
-    "lib/renderer/ipc-renderer-internal.ts",
     "package.json",
     "tsconfig.electron.json",
     "tsconfig.json",
@@ -266,7 +264,6 @@ auto_filenames = {
   ]
 
   renderer_bundle_deps = [
-    "lib/common/api/clipboard.ts",
     "lib/common/api/module-list.ts",
     "lib/common/api/native-image.ts",
     "lib/common/api/shell.ts",
@@ -276,6 +273,7 @@ auto_filenames = {
     "lib/common/reset-search-paths.ts",
     "lib/common/web-view-methods.ts",
     "lib/common/webpack-provider.ts",
+    "lib/renderer/api/clipboard.ts",
     "lib/renderer/api/context-bridge.ts",
     "lib/renderer/api/crash-reporter.ts",
     "lib/renderer/api/exports/electron.ts",
@@ -304,7 +302,6 @@ auto_filenames = {
   ]
 
   worker_bundle_deps = [
-    "lib/common/api/clipboard.ts",
     "lib/common/api/module-list.ts",
     "lib/common/api/native-image.ts",
     "lib/common/api/shell.ts",
@@ -313,6 +310,7 @@ auto_filenames = {
     "lib/common/ipc-messages.ts",
     "lib/common/reset-search-paths.ts",
     "lib/common/webpack-provider.ts",
+    "lib/renderer/api/clipboard.ts",
     "lib/renderer/api/context-bridge.ts",
     "lib/renderer/api/crash-reporter.ts",
     "lib/renderer/api/exports/electron.ts",
