+++ lib/browser/api/clipboard.ts
@@ -1,0 +1,3 @@
+const clipboard = process._linkedBinding('electron_common_clipboard');
+
+export default clipboard;
