+++ lib/browser/api/module-list.ts
@@ -7,6 +7,7 @@ export const browserModuleList: ElectronInternal.ModuleEntry[] = [
   { name: 'BaseWindow', loader: () => require('./base-window') },
   { name: 'BrowserView', loader: () => require('./browser-view') },
   { name: 'BrowserWindow', loader: () => require('./browser-window') },
+  { name: 'clipboard', loader: () => require('./clipboard') },
   { name: 'contentTracing', loader: () => require('./content-tracing') },
   { name: 'crashReporter', loader: () => require('./crash-reporter') },
   { name: 'dialog', loader: () => require('./dialog') },
